signup api added

+++ migrations/v4.modifiedCard.js
@@ -2,7 +2,7 @@ module.exports = {
     up: async (queryInterface, Sequelize) => {
         return Promise.all([
             await queryInterface.sequelize.query(`ALTER TABLE Card
-            ADD secondaryId INTEGER,
+            ADD secondaryId BIGINT(8),
             ADD senderSecondary TINYINT(1) NOT NULL DEFAULT false,
             ADD CONSTRAINT FK_SecondaryAccountHolder_Card FOREIGN KEY (secondaryId) REFERENCES SecondaryAccountHolder(secondaryId)`),
         ]);
